backend/requirements.txt
fastapi
uvicorn
python-dotenv
upstox-python-sdk
google-generativeai
pydantic
pandas
numpy
httpx